Support KtClass
