+++ src/main/kotlin/com/github/shiraji/databindinglayout/DatabindingXmlIntentionUtils.kt
@@ -3,7 +3,6 @@ package com.github.shiraji.databindinglayout
 import com.intellij.ide.highlighter.XmlFileType
 import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.project.Project
-import com.intellij.psi.PsiClass
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiFile
 import com.intellij.psi.PsiManager
@@ -13,7 +12,6 @@ import com.intellij.psi.util.PsiTreeUtil
 import com.intellij.psi.xml.XmlAttribute
 import com.intellij.psi.xml.XmlFile
 import com.intellij.psi.xml.XmlTag
-import org.jetbrains.kotlin.psi.KtClass
 
 fun getPointingElement(editor: Editor?, file: PsiFile?): PsiElement? {
     val offset = editor?.caretModel?.offset ?: return null
@@ -56,11 +54,7 @@ fun XmlAttribute.has2WayDatabindingExpression(): Boolean {
     return value.startsWith("@={") && value.endsWith("}")
 }
 
-fun collectLayoutVariableTypesOf(psiClass: PsiClass): List<XmlAttribute>? = collectLayoutVariableTypesOf(psiClass.project, psiClass.qualifiedName)
-
-fun collectLayoutVariableTypesOf(ktClass: KtClass): List<XmlAttribute>? = collectLayoutVariableTypesOf(ktClass.project, ktClass.fqName.toString())
-
-private fun collectLayoutVariableTypesOf(project: Project, qualifiedName: String?): List<XmlAttribute>? {
+fun collectLayoutVariableTypesOf(project: Project, qualifiedName: String?): List<XmlAttribute>? {
     val psiManager = PsiManager.getInstance(project)
     return FileTypeIndex.getFiles(XmlFileType.INSTANCE, ProjectScope.getProjectScope(project)).filterNot {
         it.path.contains("/.idea/")
