+++ src/main/resources/META-INF/plugin.xml
@@ -35,7 +35,7 @@
     <!-- uncomment to enable plugin in all products
     <depends>com.intellij.modules.lang</depends>
     -->
-    <depends>org.jetbrains.kotlin</depends>
+    <depends optional="true">org.jetbrains.kotlin</depends>
 
     <extensions defaultExtensionNs="com.intellij">
         <intentionAction>
@@ -79,7 +79,8 @@
         </intentionAction>
 
         <codeInsight.lineMarkerProvider language="JAVA" implementationClass="com.github.shiraji.databindinglayout.linemaker.LayoutNavigationLineMarkerProvider" />
-        <codeInsight.lineMarkerProvider language="kotlin" implementationClass="com.github.shiraji.databindinglayout.linemaker.LayoutNavigationLineMarkerProvider" />
+        <codeInsight.lineMarkerProvider language="kotlin"
+                                        implementationClass="com.github.shiraji.databindinglayout.linemaker.KtLayoutNavigationLineMarkerProvider"/>
     </extensions>
 
     <application-components>
